pyproject.toml
[project]
name = "fin-agents"
version = "0.1.0"
description = "Financial analysis agents with real-time Telegram alerts"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27,<1",
    "pdfplumber>=0.11.9",
    "python-dotenv>=1.0",
    "yfinance>=0.2",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio>=0.24", "ruff", "mypy"]

[project.scripts]
app = "app.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

# ---------- Testing ----------
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short --strict-markers"
asyncio_mode = "auto"

# ---------- Linting & Formatting ----------
[tool.ruff]
target-version = "py312"
src = ["src", "tests"]
line-length = 88

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "S",    # flake8-bandit (security)
    "A",    # flake8-builtins
    "C4",   # flake8-comprehensions
    "DTZ",  # flake8-datetimez
    "T20",  # flake8-print
    "RET",  # flake8-return
    "PTH",  # flake8-use-pathlib
    "ERA",  # eradicate (commented-out code)
    "RUF",  # ruff-specific rules
]
ignore = [
    "S101",  # allow assert in tests
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "S105", "S106", "T20"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

# ---------- Type Checking ----------
[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
mypy_path = "src"

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = ["yfinance", "yfinance.*"]
ignore_missing_imports = true
